feat: config watcher
