+++ src/templates/typeScriptFiles.ts
@@ -3,7 +3,7 @@ function capitalizeFirstLettee(name: string) {
         .replace(/\/([a-z])/g, (g) => g[1].toUpperCase())
 }
 
-export const apiTemplate = (name: string) => `export default defineEventHandler((event) => {
+export const apiTemplate = (name: string) => `export default defineEventHandler(async (event) => {
   return 'Hello ${name}'
 })
 `
